Kubernetes_part/kubernetes_postgres/pub_kubernetes/requirements.txt
Django
python-dotenv
psycopg2-binary
requests
django-cors-headers